Conflicts:
	src/main/java/com/softserve/edu/Resources/controller/LookUpController.java
	src/main/java/com/softserve/edu/Resources/entity/PropertyValue.java
	src/main/java/com/softserve/edu/Resources/service/impl/ResourceServiceImpl.java

+++ src/main/java/com/softserve/edu/Resources/dto/ResourceTypeDTO.java
@@ -7,11 +7,9 @@ public class ResourceTypeDTO {
     private String tableName;
     
     public ResourceTypeDTO() {
-       
     }
     
-    public ResourceTypeDTO(int id, String typeName, String tableName) {
-        
+    public ResourceTypeDTO(Long id, String typeName, String tableName) {
         this.id = id;
         this.typeName = typeName;
         this.tableName = tableName;
@@ -20,18 +18,23 @@ public class ResourceTypeDTO {
     public long getId() {
         return id;
     }
+
     public void setId(long id) {
         this.id = id;
     }
+
     public String getTypeName() {
         return typeName;
     }
+
     public void setTypeName(String typeName) {
         this.typeName = typeName;
     }
+
     public String getTableName() {
         return tableName;
     }
+
     public void setTableName(String tableName) {
         this.tableName = tableName;
     }
